+++ app/AppKernel.php
@@ -21,7 +21,7 @@ class AppKernel extends Kernel
             new Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
             new Sensio\Bundle\FrameworkExtraBundle\SensioFrameworkExtraBundle(),
             new Knp\Bundle\PaginatorBundle\KnpPaginatorBundle(),
-            new ShowSourceBundle\ShowSourceBundle(),
+            new CodeExplorerBundle\CodeExplorerBundle(),
             new AppBundle\AppBundle(),
         );
