--- src/ShowSourceBundle/Resources/config/services.yml
@@ -1,13 +1,0 @@
-services:
-    show_source.twig.source_code_extension:
-        public:    false
-        class:     ShowSourceBundle\Twig\SourceCodeExtension
-        arguments: ["@twig.loader", "%kernel.root_dir%"]
-        tags:
-            - { name: twig.extension }
-
-    show_source.controller_listener:
-        class:     ShowSourceBundle\EventListener\ControllerListener
-        arguments: ["@show_source.twig.source_code_extension"]
-        tags:
-            - { name: 'kernel.event_listener', event: kernel.controller, method: registerCurrentController }
